smaller ids

+++ test/Positions.t.sol
@@ -16,12 +16,12 @@ contract PositionsTest is FullTest {
 
     function test_saltToId(address minter, bytes32 salt) public {
         uint256 id = positions.saltToId(minter, salt);
-        assertLe(id, type(uint64).max);
+        assertLe(id, type(uint48).max);
         unchecked {
             assertNotEq(id, positions.saltToId(address(uint160(minter) + 1), salt));
             assertNotEq(id, positions.saltToId(minter, bytes32(uint256(salt) + 1)));
         }
-        // address is also incorporatedll
+        // address is also incorporated
         Positions p2 = new Positions(core, positions.tokenURIGenerator());
         assertNotEq(id, p2.saltToId(minter, salt));
     }
